Moved Routers into api folder
